Force enabled printf format checking in project.yml in order to catch issues that GCC clang is not catching with mismatched specifiers. Ensured that FILE* is cast to size_t to eliminate platform specific complaints. Readded sleep to teardown in socket tests, since I guess it DOES matter on Linux... UGH

+++ src/lib/kinetic_logger.c
@@ -41,7 +41,7 @@ void KineticLogger_Init(const char* logFile)
         }
         else
         {
-            fprintf(stderr, "Failed to initialize logger with file: fopen('%s') => fd=%zd", logFile, fd);
+            fprintf(stderr, "Failed to initialize logger with file: fopen('%s') => fd=%zd", logFile, (size_t)fd);
             fflush(fd);
         }
     }
